fix wrong pid issue

+++ ansible/roles/codebase_dr/files/DRWorker/dr_worker_service.py
@@ -4,13 +4,14 @@ import time
 from subprocess import Popen
 from subprocess import PIPE
 from subprocess import STDOUT
+import traceback as tb
 
 def launchCore(campaign,campaign_datasource):
   if campaign_datasource[4] == 'dataset':
     _cmd = ['./process_dataset.sh',campaign[1],campaign[4],campaign_datasource[3],campaign_datasource[2]]
     pid = Popen(_cmd, stdout=PIPE, stderr=STDOUT, stdin=PIPE)
     db=DRDB("/var/local/LNEx.db")
-    db.add_drworker(campaign[1],campaign_datasource[2],pid)
+    db.add_drworker(campaign[1],campaign_datasource[2],pid.pid)
     db.update_media_object_status(campaign_datasource[2],1)
     db.destroy_connection()
 
@@ -27,14 +28,18 @@ def launchCore(campaign,campaign_datasource):
 
 def checkStatus(campaign_datasource):
   db=DRDB("/var/local/LNEx.db")
+  print(campaign_datasource[1], campaign_datasource[2])
   drworkerEntry=db.grab_drworker(campaign_datasource[1], campaign_datasource[2])
   db.destroy_connection()
-  try:
-    os.kill(drworkerEntry[3], 0)
-  except OSError:
-    db=DRDB("/var/local/LNEx.db")
-    db.update_media_object_status(campaign_datasource[2], -1)
-    db.destroy_connection()
+  if len(drworkerEntry) > 0:
+    try:
+      os.kill(int(drworkerEntry[0][3]), 0)
+    except OSError:
+      var = tb.format_exc()
+      print(var)
+      db=DRDB("/var/local/LNEx.db")
+      db.update_media_object_status(campaign_datasource[2], -1)
+      db.destroy_connection()
 
 while True:
   db=DRDB("/var/local/LNEx.db")
